Introduce code to replace dependency on clapack
This code re-implements functionality that is normally provided through a
lapack implementation. It was undesirable to include an additional
dependency in the LSST stack, and because of the poor structure of the
psfex codebase it was easier to create a library that pretended to be a
full lapack implementation and ship it with psfex. Because it is not a full
implementation (it only implements two functions) it was best not to ship
it as an additional LSST package, so no future developer would mistake the
functionality.