Turned off plplot from build
The plplot functionality psfex uses is missing from the latest builds of
plplot, and will cause the build to fail on systems with plplot. Because we
do not use the plotting functionality in the LSST stack, it is being
disabled explicitly (instead of assuming it is not present on a system) to
prevent failures.