Be more explicit about initializing variables.
Avoid "warning: 'end' may be used uninitialized in this function" output from
the compiler by initializing 'end' and 'start' at the same time. The warning
was misleading -- 'end' was, in fact, always initialized correctly -- but the
compiler isn't smart enough to figure that out so it's better to suppress the
warning.