Fix resampling of basis functions
PSFEx resamples the basis functions of the PSF model to shift the model
to the center of a star using Lanczos interpolation.  It assumes that
these basis function are delta functions, but these are not well sampled
and thus introduce aliasing.

The new function (vignet_resample_pixel) shifts the PSF basis functions by
calculating a pixel overlap integral with the native pixel scale.